backend/requirements.txt
flask
flask-cors
google-generativeai
nba_api
pandas
rapidfuzz
python-dotenv
